Move `testkit` under test
This change places `testkit` where it belongs, avoiding shipping it
with the rest of the codebase.

This might seem a trivial, but binary analysis tools might flag
ToolHive binaries as insecure because of that.
