+++ pkg/mcp/tool_middleware_test.go
@@ -9,7 +9,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/stacklok/toolhive/pkg/testkit"
+	"github.com/stacklok/toolhive/test/testkit"
 )
 
 func TestNewListToolsMappingMiddleware_Scenarios(t *testing.T) {
