+++ test/testkit/sse_server.go
